Updated UserController, UserRepository, UserService and .properties. Created ExceptionHandler, Utils.java and PlanService
Updated UserController, implemented methods to sign up and sign in, and another to log in with a cookie session.

Updated UserSession to bring all the logic to Sign up and Sign in forms.

Created UniqueValueException to be used with unique values in DB.

Created PlanService to apply to the new user a existant plan.

Updated UserRepositoty with more queries.

Created Utils.java to save utils methods such convert camelCase to normal text...