Created existsUserWith methods, ImageService  and updateUser method
Added three variants of existsUserWith, one with username, phone number and email. (in controller and service).

Created ImageStorageService, that brings the possibility to upload image files (that can be used to change our profile picture), the service moves the files into an external folder and adds it a UUID.

In UserController, added a new method to update the user profile, it need a user in json as text to be converted, a multipartfile (can be null) and the session UUID to grant permissions to the user to modify his own profile only.

Added a config file, that contains a resource handler, to track the image folder and serve its content.