Added FollowerController and FollowerService
Added FollowerController to allow API connections to query followers, followings and add new followings from a given user to another.(it is only possible to make a new following if the session cookie is present and has a valid UUID).

Added FollowersService, which queries and makes new follows for given users.

Fixed some bugs with loops when serializing to json in User class.