Added some personalized querys to repositories
In NoteRepository, we created a "deleteAllByPanel" where you pass as a paramenter a panel object and all notes that are in the panel will disappear.

In  PanelParticipantRepository, we added "findPanelParticipantByParticipantAndPAnel" where if a user and a panel are passed, and theres any PanelParticipant in that panel with that user, it will be returned.

and, also, "findPanelPariticipantsByPanel" where a list of all participants of that panel will be returned.

In UserRepository, we aded "findUserByNicknameContainingIgnoreCase", where it will search any user (and return a list of them), that contains the search string passed as a parameter ignoring the cases.