Added PanelParticipantsController and some little changes
PanelParticipantsController have been agregated, with some end points to consult useful information to show in web, on the panels section, among that changes, are: add an delete users by an array of user ids.

Added to the panel class, a new field, to save if a panel is blocked.

Added a constructor for PanelParticipant class and PanelParticipantId.

Implemented all the functionalities to the PanelParticipantsService needed in the correspond service.