Added methods to user and panel services and controllers to get100 random results
Added the method "find100" in user and panel controller to get 100 random results to show them in the "Explore" page in the web side.

Added equals and HasCode in user and panel model class.

Added custom querys in user and panel repositories to get the number of the las id in the table.